requirements.txt

pillow
random
psutil
requests
bs4
json